fixes as per the comments
