+++ mm2src/coins/utxo/qtum.rs
@@ -3,7 +3,6 @@ use crate::init_withdraw::{InitWithdrawCoin, WithdrawTaskHandle};
 use crate::{eth, CanRefundHtlc, CoinBalance, DelegationError, DelegationFut, GetRawTransactionError,
             NegotiateSwapContractAddrErr, StakingInfosFut, SwapOps, TradePreimageValue, ValidateAddressResult,
             WithdrawFut};
-use bitcoin_hashes::hex::ToHex;
 use common::mm_metrics::MetricsArc;
 use common::mm_number::MmNumber;
 use crypto::trezor::TrezorCoin;
@@ -622,32 +621,8 @@ impl MarketCoinOps for QtumCoin {
         utxo_common::send_raw_tx(&self.utxo_arc, tx)
     }
 
-    fn get_raw_tx(
-        &self,
-        mut tx: &str,
-    ) -> Box<dyn Future<Item = String, Error = MmError<GetRawTransactionError>> + Send> {
-        if tx.starts_with("0x") {
-            tx = &tx[2..];
-        }
-        match H256Json::from_str(tx) {
-            Ok(tx_hash) => {
-                let qtum_coin_fields = self.utxo_arc.clone();
-                Box::new(
-                    Box::pin(async move {
-                        qtum_coin_fields
-                            .rpc_client
-                            .get_rpc_transaction(&tx_hash)
-                            .await
-                            .map(|raw_tx| raw_tx.hex.to_hex())
-                            .map_err(|err| err.map(GetRawTransactionError::from))
-                    })
-                    .compat(),
-                )
-            },
-            Err(err) => Box::new(futures01::future::err(
-                GetRawTransactionError::InvalidTxHash(err.to_string()).into(),
-            )),
-        }
+    fn get_raw_tx(&self, tx: &str) -> Box<dyn Future<Item = String, Error = MmError<GetRawTransactionError>> + Send> {
+        utxo_common::get_raw_tx(self.utxo_arc.rpc_client.clone(), tx)
     }
 
     fn wait_for_confirmations(
